emarketproject/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==22.12.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
crispy-bootstrap5==0.7
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.3
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-ckeditor-updated==4.4.4
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.3.0
django-extensions==3.1.3
django-filter==23.5
django-jazzmin==2.6.0
django-jet-reboot==1.3.5
django-js-asset==2.2.0
django-model-utils==4.3.1
django-paypal==2.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
django-user-visit==2.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
gopay==2.0.1
idna==3.6
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
Pillow==9.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.14
Pygments==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7
pyzmq==25.1.2
razorpay==1.4.1
reportlab==4.0.4
requests==2.31.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==5.5.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
wcwidth==0.2.13
whitenoise==6.2.0
wrapt==1.16.0
